pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fail2ban-telegram-bot"
version = "1.0.0"
description = "Telegram bot for monitoring fail2ban: real-time stats, service status, and global geo-mapping of banned IPs. Visualize attack sources on interactive-style world maps, track bans over time, and get notified when your firewall blocks threats — all through Telegram. Docker-ready and easy to deploy."
authors = [
  { name="ksalab", email="ksalab0@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# Core dependencies for the bot to run
dependencies = [
    "aiogram==3.10.0",
    "python-telegram-bot==20.6",
    "pydantic-settings==2.3.4",
    "python-dotenv==1.0.1",
    "python-dateutil==2.9.0",
    "geoip2==4.8.0",
    "aiohttp==3.9.5",
    "matplotlib==3.9.0",
    "pandas==2.2.2",
    "cartopy==0.23.0",
    "Shapely==2.1.1",
]

[project.urls]
"Homepage" = "https://github.com/ksalab/fail2ban_bot" # Optional: Link to your repo
"Bug Tracker" = "https://github.com/ksalab/fail2ban_bot/issues" # Optional

# Optional dependencies for development (e.g., linters, formatters)
[project.optional-dependencies]
dev = [
    "ruff",
    "black",
]
